+++ packages/storage/src/length_prefixed.rs
@@ -29,6 +29,14 @@ pub fn to_length_prefixed_nested(namespaces: &[&[u8]]) -> Vec<u8> {
     out
 }
 
+pub fn length_prefixed_with_key(namespace: &[u8], key: &[u8]) -> Vec<u8> {
+    let mut out = Vec::with_capacity(namespace.len() + 2 + key.len());
+    out.extend_from_slice(&encode_length(namespace));
+    out.extend_from_slice(namespace);
+    out.extend_from_slice(key);
+    out
+}
+
 /// This is equivalent concat(to_length_prefixed_nested(namespaces), key)
 /// But more efficient when the intermediate namespaces often must be recalculated
 #[allow(dead_code)]
@@ -47,6 +55,30 @@ pub fn namespaces_with_key(namespaces: &[&[u8]], key: &[u8]) -> Vec<u8> {
     out
 }
 
+/// Customization of namespaces_with_key for when
+/// there are multiple sets we do not want to combine just to call this
+pub fn nested_namespaces_with_key(top_names: &[&[u8]], sub_names: &[&[u8]], key: &[u8]) -> Vec<u8> {
+    let mut size = key.len();
+    for &namespace in top_names {
+        size += namespace.len() + 2;
+    }
+    for &namespace in sub_names {
+        size += namespace.len() + 2;
+    }
+
+    let mut out = Vec::with_capacity(size);
+    for &namespace in top_names {
+        out.extend_from_slice(&encode_length(namespace));
+        out.extend_from_slice(namespace);
+    }
+    for &namespace in sub_names {
+        out.extend_from_slice(&encode_length(namespace));
+        out.extend_from_slice(namespace);
+    }
+    out.extend_from_slice(key);
+    out
+}
+
 /// Encodes the length of a given namespace as a 2 byte big endian encoded integer
 fn encode_length(namespace: &[u8]) -> [u8; 2] {
     if namespace.len() > 0xFFFF {
@@ -54,6 +86,12 @@ fn encode_length(namespace: &[u8]) -> [u8; 2] {
     }
     let length_bytes = (namespace.len() as u32).to_be_bytes();
     [length_bytes[2], length_bytes[3]]
+}
+
+// pub(crate) fn decode_length(prefix: [u8; 2]) -> usize {
+pub(crate) fn decode_length(prefix: &[u8]) -> usize {
+    // TODO: enforce exactly 2 bytes somehow, but usable with slices
+    (prefix[0] as usize) * 256 + (prefix[1] as usize)
 }
 
 #[cfg(test)]